geofence클래스 타입오류 해결 / 맵 위치 수정 후 geofence 변경 추가

+++ app/src/main/java/org/konkuk/placelist/Geofence.kt
@@ -35,7 +35,7 @@ class Geofence(private val context: Context) {
             geofenceList.add(
                 getGeofence(
                     now.name,
-                    LatLng(now.latitude, now.longitude),
+                    LatLng(now.latitude.toDouble(), now.longitude.toDouble()),
                     now.detectRange
                 )
             )
